# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateSubscriptionRequest:
    properties:
      activeUntil:
        type: string
      planId:
        type: integer
      userId:
        type: integer
    required:
    - planId
    - userId
    type: object
...
  dto.SubscriptionPlanResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      durationDays:
        type: integer
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: integer
      updatedAt:
        type: string
    type: object
  dto.SubscriptionResponse:
    properties:
      activeUntil:
        type: string
      cancelledAt:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      plan:
        $ref: '#/definitions/dto.SubscriptionPlanResponse'
      planId:
        type: integer
      status:
        $ref: '#/definitions/models.SubscriptionStatus'
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
...
  dto.UpdateSubscriptionRequest:
    properties:
      activeUntil:
        type: string
      cancelledAt:
        type: string
      planId:
        type: integer
      status:
        $ref: '#/definitions/models.SubscriptionStatus'
    type: object
...
  models.SubscriptionStatus:
    enum:
    - active
    - expired
    - cancelled
    type: string
    x-enum-varnames:
    - SubscriptionStatusActive
    - SubscriptionStatusExpired
    - SubscriptionStatusCancelled
...
host: localhost:8080
info:
  contact: {}
  description: Api server for DaraTil PLatform
  title: DaraTil Backend
  version: "0.2"
paths:
...
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Get subscriptions with filters
      parameters:
      - description: User ID
        in: query
        name: userId
        type: integer
      - description: Plan ID
        in: query
        name: planId
        type: integer
      - description: Subscription status
        in: query
        name: status
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SubscriptionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List subscriptions
      tags:
      - Subscriptions
    post:
      consumes:
      - application/json
      description: Create a new user subscription
      parameters:
      - description: Subscription create request
        in: body
        name: subscriptionReq
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create subscription
      tags:
      - Subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete subscription
      tags:
      - Subscriptions
    get:
      consumes:
      - application/json
      description: Get subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get subscription by ID
      tags:
      - Subscriptions
    patch:
      consumes:
      - application/json
      description: Update subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subscription update request
        in: body
        name: subscriptionReq
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update subscription
      tags:
      - Subscriptions
  /subscriptions/{id}/cancel:
    patch:
      consumes:
      - application/json
      description: Cancel subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel subscription
      tags:
      - Subscriptions
  /subscriptions/{id}/expire:
    patch:
      consumes:
      - application/json
      description: Mark subscription as expired by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Expire subscription
      tags:
      - Subscriptions
  /subscriptions/users/{id}/active:
    get:
      consumes:
      - application/json
      description: Get active subscription of user by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get active subscription by user ID
      tags:
      - Subscriptions
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"